telegram-bot-service/requirements.txt
aiogram==3.0.0
asyncpg
SQLAlchemy
pydantic
requests
python-dotenv
